feat: enhance repayment and savings components with improved mobile responsiveness and color coding
- Updated RentTable to split items into left and right columns for better mobile display.
- Changed month display format from short to long in RentTable and other components.
- Introduced category color coding for repayments and transactions for better visual distinction.
- Enhanced DailyRepaymentChart and SavingsCharts to adjust bar sizes based on mobile view.
- Improved layout and styling in RepaymentPivot, RepaymentTable, and SavingsTable for mobile compatibility.
- Added new category source "barclays" in TransactionTable and TransactionsTab.
- Refactored chart tick rendering for better spacing and clarity.
- Updated formatMonthLabel to use full month names for consistency across the application.

+++ src/lib/format.ts
@@ -43,9 +43,9 @@ export function recentMonths(count = 18): string[] {
   return months;
 }
 
-/** Format "2026-06" → "Jun 2026" */
+/** Format "2026-06" → "June 2026" (full month name — used everywhere except graphs) */
 export const formatMonthLabel = (m: string): string => {
   const [year, month] = m.split("-");
   const d = new Date(parseInt(year), parseInt(month) - 1, 1);
-  return `${d.toLocaleString("en-GB", { month: "short" })} ${year}`;
+  return `${d.toLocaleString("en-GB", { month: "long" })} ${year}`;
 };
